+++ src/base/gfx/colorgradient.h
@@ -17,8 +17,8 @@ struct ColorGradient : Math::SegmentedFunction<Color, ColorGradient>
 
 	using SegmentedFunction::SegmentedFunction;
 
-	explicit ColorGradient(
-	    const std::string &stoplist = std::string());
+	[[nodiscard]] static ColorGradient fromString(
+	    const std::string &stoplist);
 
 	explicit operator std::string() const;
 };